requirements.txt
absl-py==1.4.0
albumentations==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
black==23.7.0
bleach==6.0.0
cachetools==5.3.0
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.3
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-jasmine==0.5.1
django-jazzmin==2.4.8
djangorestframework==3.14.0
ecg-plot==0.2.8
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.2
fqdn==1.5.1
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.4.0
google-ai-generativelanguage==0.1.0
google-api-core==2.13.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
grpcio==1.59.2
grpcio-status==1.59.2
h5py==3.8.0
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==15.0.6.1
lit==16.0.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.0.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
pgmpy==0.1.22
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
plotly==5.14.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.19.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
SimpleITK==2.3.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
tk==0.1.0
tkintertable==1.3.3
tomli==2.0.1
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
zipp==3.15.0